Moved management client and api classes into separate package
