+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/management/ManagementIncidentEntityUpdater.java
@@ -19,6 +19,7 @@ import fi.otavanopisto.kuntaapi.server.discover.EntityUpdater;
 import fi.otavanopisto.kuntaapi.server.id.IncidentId;
 import fi.otavanopisto.kuntaapi.server.id.OrganizationId;
 import fi.otavanopisto.kuntaapi.server.integrations.KuntaApiConsts;
+import fi.otavanopisto.kuntaapi.server.integrations.management.client.ManagementApi;
 import fi.otavanopisto.kuntaapi.server.integrations.management.resources.ManagementIncidentResourceContainer;
 import fi.otavanopisto.kuntaapi.server.integrations.management.tasks.IncidentIdTaskQueue;
 import fi.otavanopisto.kuntaapi.server.persistence.model.Identifier;
