+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/management/ManagementPageIdMapEntityUpdater.java
@@ -26,6 +26,7 @@ import fi.otavanopisto.kuntaapi.server.id.OrganizationId;
 import fi.otavanopisto.kuntaapi.server.id.PageId;
 import fi.otavanopisto.kuntaapi.server.integrations.IdMapProvider.OrganizationPageMap;
 import fi.otavanopisto.kuntaapi.server.integrations.KuntaApiConsts;
+import fi.otavanopisto.kuntaapi.server.integrations.management.client.ManagementApi;
 import fi.otavanopisto.kuntaapi.server.integrations.management.resources.ManagementPageIdMapResourceContainer;
 import fi.otavanopisto.kuntaapi.server.integrations.management.tasks.OrganizationPageMapsTaskQueue;
 import fi.otavanopisto.kuntaapi.server.settings.OrganizationSettingController;
